Fix bug that caused the app to crush in some API levels e.g API level 24.

+++ feature/artistdetails/src/main/java/com/techdroidcentre/artistdetails/ArtistDetailsUiState.kt
@@ -25,7 +25,7 @@ fun MediaItem.toAlbum(): Album {
 
 fun MediaItem.toArtist(): Artist {
     return Artist(
-        id = mediaId.toLong(),
+        id = mediaId,
         uri = localConfiguration?.uri.toString(),
         name = mediaMetadata.title.toString()
     )
